Update thumbnail.vue
The .thumbnail class overlaps other .thumbnail classes.
For example if you already have, for example, a bootstrap thumbnail element in your site, the class will change the element, since it is not bound to any prior class. 

The .thumbnail class should be unique to the component.

Changed the class name.